feat(cli): introduce commander framework with diagnose/init/dev subcommands
- Add commander dep; register both `3amoncall` and `3amoncall-cli` bins
- Move run() logic to commands/diagnose.ts (runDiagnose)
- Add stub commands/init.ts and commands/dev.ts (Task 3/5 implement)
- index.ts delegates run() → runDiagnose for backward compat
- All 9 existing tests pass
