+++ packages/cli/package.json
@@ -6,6 +6,7 @@
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
   "bin": {
+    "3amoncall": "./dist/cli.js",
     "3amoncall-cli": "./dist/index.js"
   },
   "exports": {
@@ -22,7 +23,8 @@
   },
   "dependencies": {
     "@3amoncall/core": "workspace:*",
-    "@3amoncall/diagnosis": "workspace:*"
+    "@3amoncall/diagnosis": "workspace:*",
+    "commander": "^14.0.3"
   },
   "devDependencies": {
     "@3amoncall/config-eslint": "workspace:*",
