+++ packages/cli/src/commands/dev.ts
@@ -1,0 +1,4 @@
+export async function runDev(_argv: string[]): Promise<void> {
+  process.stderr.write("Error: 'dev' command is not yet implemented\n");
+  process.exit(1);
+}
